backend/requirements.txt
attrs==25.3.0
basemap==1.4.1
basemap-data==1.3.2
blinker==1.9.0
branca==0.8.1
Cartopy==0.24.1
certifi==2024.12.14
cffi==1.17.1
cfgrib==0.9.15.0
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
eccodes==2.41.0
findlibs==0.1.1
Flask==3.1.2
flask-cors==6.0.2
folium==0.19.4
fonttools==4.55.7
glcontext==2.5.0
h5netcdf==1.6.1
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
ManimPango==0.5.0
MarkupSafe==3.0.3
matplotlib==3.10.0
mdurl==0.1.2
moderngl==5.10.0
multipledispatch==1.0.0
netCDF4==1.7.2
networkx==3.3
numpy==1.26.4
packaging==23.2
pandas==2.2.3
pillow==10.3.0
py-cpuinfo==9.0.0
pycairo==1.26.0
pycparser==2.22
pydub==0.25.1
pyglet==2.0.15
Pygments==2.18.0
pyparsing==3.2.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scipy==1.13.0
screeninfo==0.8.1
shapely==2.0.7
six==1.17.0
skia-pathops==0.8.0.post1
srt==3.5.3
svgelements==1.9.6
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2025.1
urllib3==2.3.0
watchdog==4.0.0
Werkzeug==3.1.5
xarray==2025.6.1
xyzservices==2025.1.0
python-docx
docx2txt
regex
chardet
pysmb

